Move filter functions to app component
